+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project uses [Semantic Versioning](http://semver.org/).
 
+# [0.1.0] - 2020-9-8
+### Added
+ - `Deal_or_No_Deal` environment in `envs`
+ - `BankerModel` in `model.py`
+ - `DQNAgent` code in `dqn.py`
+
 # [0.0.1] - 2020-8-29
 ### Added
  - boilerplate code format for OpenAI's `gym`